requirements.txt
Django~=5.1.2
djangorestframework~=3.15.2
Pillow
django-model-utils~=5.0.0
django-filter
